refactor(templates): strip TypeScript — just markdown files
The templates package is a directory of .md files, not a TypeScript
module. Consumers read the files they need directly. No build step,
no type declarations, no index.ts wrapper. Works for any language —
TypeScript, Go, whatever comes next.

CLI awaken.ts now reads awaken.md via readFileSync with its own
renderTemplate helper (same pattern as hooks.ts).
