+++ packages/cli/package.json
@@ -8,7 +8,7 @@
     "sb": "dist/cli.js"
   },
   "scripts": {
-    "build": "yarn workspace @personal-context/templates build && tsc && cp -r src/templates dist/templates",
+    "build": "tsc && cp -r src/templates dist/templates",
     "dev": "tsc --watch",
     "cli": "tsx src/cli.ts",
     "install:cli": "chmod +x dist/cli.js && mkdir -p \"$HOME/.local/bin\" && ln -sf \"$(pwd)/dist/cli.js\" \"$HOME/.local/bin/sb\" && echo 'Linked: $HOME/.local/bin/sb → dist/cli.js'",
@@ -19,7 +19,6 @@
   },
   "dependencies": {
     "@inquirer/prompts": "^8.2.0",
-    "@personal-context/templates": "*",
     "chalk": "^5.3.0",
     "commander": "^12.1.0",
     "ora": "^8.0.1"
