+++ packages/templates/awaken.md
