+++ packages/templates/starters/soul.md
