+++ packages/templates/starters/values.md
